# Make urlutil.Domain 10x faster