--- a/server/util/urlutil/urlutil_test.go
+++ b/server/util/urlutil/urlutil_test.go
@@ -11,10 +11,30 @@ func TestGetDomain(t *testing.T) {
 	for _, test := range []struct {
 		host, domain string
 	}{
+		{host: "buildbuddy.io", domain: "buildbuddy.io"},
 		{host: "app.buildbuddy.io", domain: "buildbuddy.io"},
+		{host: "foo.app.buildbuddy.io", domain: "buildbuddy.io"},
 		{host: "app.localhost.io:8080", domain: "localhost.io"},
+		{host: "localhost:8080", domain: "localhost"},
 	} {
 		domain := urlutil.GetDomain(test.host)
 		assert.Equal(t, test.domain, domain, "GetDomain(%q)", test.host)
+	}
+}
+
+func BenchmarkGetDomain(b *testing.B) {
+	for _, test := range []struct {
+		host, domain string
+	}{
+		{host: "buildbuddy.io", domain: "buildbuddy.io"},
+		{host: "app.buildbuddy.io", domain: "buildbuddy.io"},
+		{host: "foo.app.buildbuddy.io", domain: "buildbuddy.io"},
+		{host: "app.localhost.io:8080", domain: "localhost.io"},
+	} {
+		b.Run(test.host, func(b *testing.B) {
+			for b.Loop() {
+				_ = urlutil.GetDomain(test.host)
+			}
+		})
 	}
 }