--- a/server/util/urlutil/urlutil_test.go
+++ b/server/util/urlutil/urlutil_test.go
@@ -30,6 +30,7 @@ func BenchmarkGetDomain(b *testing.B) {
 		{host: "app.buildbuddy.io", domain: "buildbuddy.io"},
 		{host: "foo.app.buildbuddy.io", domain: "buildbuddy.io"},
 		{host: "app.localhost.io:8080", domain: "localhost.io"},
+		{host: "localhost:8080", domain: "localhost"},
 	} {
 		b.Run(test.host, func(b *testing.B) {
 			for b.Loop() {