It's currently 0.5% of app CPU.

Benchmark results:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor
                                   │  domainbase   │            domainindex             │
                                   │    sec/op     │   sec/op     vs base               │
GetDomain/buildbuddy.io-24           68.030n ± 21%   6.869n ± 1%  -89.90% (p=0.000 n=9)
GetDomain/app.buildbuddy.io-24       83.770n ± 14%   6.023n ± 3%  -92.81% (p=0.000 n=9)
GetDomain/foo.app.buildbuddy.io-24   89.880n ± 30%   5.944n ± 3%  -93.39% (p=0.000 n=9)
GetDomain/app.localhost.io:8080-24   57.740n ± 28%   8.331n ± 3%  -85.57% (p=0.000 n=9)
geomean                               73.74n         6.728n       -90.88%

                                   │ domainbase │              domainindex              │
                                   │    B/op    │   B/op     vs base                    │
GetDomain/buildbuddy.io-24           48.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=9)
GetDomain/app.buildbuddy.io-24       64.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=9)
GetDomain/foo.app.buildbuddy.io-24   80.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=9)
GetDomain/app.localhost.io:8080-24   64.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=9)
geomean                              62.98                   ?                      ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                                   │ domainbase │              domainindex               │
                                   │ allocs/op  │ allocs/op   vs base                    │
GetDomain/buildbuddy.io-24           2.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=9)
GetDomain/app.buildbuddy.io-24       2.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=9)
GetDomain/foo.app.buildbuddy.io-24   2.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=9)
GetDomain/app.localhost.io:8080-24   2.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=9)
geomean                              2.000                    ?                      ¹ ²
```